+++ src/Integrator.py
@@ -13,8 +13,13 @@ import os
 from pathlib import Path
 import time
 import torch
+from .WikidataEmbeddingGenerator import WikidataEmbeddingGenerator
+from .Matcher import Matcher
 
 class Integrator:
+    """
+    Class for integrating triples into the Wikidata Ontology.
+    """
     def __init__(self, input_dir: str, output_dir: str, embedding_model: str, matching_config: Dict = None):
         """
         Initialize the Integrator with input and output directories.
@@ -27,352 +32,71 @@ class Integrator:
         """
         self.input_dir = input_dir
         self.output_dir = output_dir
-        self.g = Graph()
-        self.wd = Namespace("http://www.wikidata.org/entity/")
-        self.wdt = Namespace("http://www.wikidata.org/prop/direct/")
-        self.g.bind("wd", self.wd)
-        self.g.bind("wdt", self.wdt)
         
+        # Initialize caches and properties
         self.entity_cache = {}
         self.property_cache = {}
-        
+        self.properties = {}
+
         # Set up matching configuration
         self.matching_config = matching_config or {
             "use_aliases": True,
             "properties_file": "wikidata-properties-with-aliases.json"
         }
 
+        # Initialize embedding model
         self.embedding_model = SentenceTransformer(embedding_model)
-        self.properties = {}
-        
-        # Load appropriate properties file based on configuration
-        self.initialize_properties()
+        self._initialize_knowledge_graph()
+        self._initialize_properties()
     
-    def initialize_properties(self):
+    def _initialize_knowledge_graph(self):
+        """Initialize the RDF graph and set up Wikidata namespaces."""
+        try:
+            self.g = Graph()
+
+            self.wd = Namespace("http://www.wikidata.org/entity/")
+            self.wdt = Namespace("http://www.wikidata.org/prop/direct/")
+            
+            self.g.bind("wd", self.wd)
+            self.g.bind("wdt", self.wdt)
+            
+            print("✅ Successfully initialized knowledge graph and namespaces")
+
+        except Exception as e:
+            print(f"❌ Error initializing knowledge graph: {str(e)}")
+            raise
+
+    def _initialize_properties(self):
         """Initialize properties based on matching configuration."""
         try:
             properties_file = self.matching_config.get("properties_file")
+            use_aliases = self.matching_config.get("use_aliases", True)
+            
+            generator = WikidataEmbeddingGenerator(embedding_model=self.embedding_model)
+            
             if not os.path.exists(properties_file):
                 print(f"Properties file {properties_file} not found. Generating...")
-                if self.matching_config.get("use_aliases", True):
-                    self.load_wikidata_properties_with_aliases(output_file=properties_file)
-                else:
-                    self.load_wikidata_properties(output_file=properties_file)
+                self.properties = generator.generate_embeddings(
+                    output_file=properties_file,
+                    use_aliases=use_aliases
+                )
+            else:
+                self.properties = generator.load_embeddings(
+                    file_path=properties_file,
+                    use_aliases=use_aliases
+                )
+
+            self.matcher = Matcher(
+                embedding_model=self.embedding_model,
+                properties=self.properties,
+                matching_config=self.matching_config
+            )
             
-            self.load_embeddings(file_path=properties_file)
             print(f"✅ Successfully initialized properties using {properties_file}")
+
         except Exception as e:
             print(f"❌ Error initializing properties: {str(e)}")
             raise
-
-    def load_wikidata_properties(self, output_file="wikidata-properties.json") -> Dict[str, Dict[str, Union[str, float, List[str]]]]:
-        """
-        Lädt alle Wikidata-Properties mit Labels, Beschreibungen, Aliassen und generiert Embeddings
-        !!! Es funktioniert zurzeit nicht ganz !
-
-        For aliases :
-        SELECT ?property ?propertyLabel ?propertyDescription ?altLabel
-        WHERE {
-        ?property a wikibase:Property .
-        SERVICE wikibase:label { 
-        bd:serviceParam wikibase:language "en" .
-        ?property rdfs:label ?propertyLabel .
-        }
-        OPTIONAL { ?property schema:description ?propertyDescription . FILTER(LANG(?propertyDescription) = "en") }
-        OPTIONAL { ?property skos:altLabel ?altLabel . FILTER(LANG(?altLabel) = "en") }
-        }
-        """
-
-        sparql = SPARQLWrapper("https://query.wikidata.org/sparql")
-        sparql.addCustomHttpHeader('User-Agent', 'Bot/1.0')
-        query = """
-        SELECT ?property ?propertyLabel
-        WHERE {
-          ?property a wikibase:Property .
-          SERVICE wikibase:label { 
-            bd:serviceParam wikibase:language "en" .
-            ?property rdfs:label ?propertyLabel .
-          }
-        }
-        """
-        
-        sparql.setQuery(query)
-        sparql.setReturnFormat(JSON)
-        sparql.setTimeout(60)
-        properties = {}
-
-        try:
-            results = sparql.query().convert()
-            
-            for result in results["results"]["bindings"]:
-                prop_id = result["property"]["value"].split("/")[-1]
-                label = result["propertyLabel"]["value"]
-
-                embedding = self.embedding_model.encode(label).tolist()
-                properties[prop_id] = {
-                    "label": label,
-                    "embedding": embedding
-                }
-            
-            with open(output_file, "w") as file:
-                json.dump(properties, file, indent=4)
-            print(f"Properties saved to {output_file}")
-        
-        except Exception as e:
-            print(f"Error fetching properties: {e}")
-
-    def load_wikidata_properties_with_aliases(self, output_file="wikidata-properties-with-aliases.json") -> Dict[str, Dict[str, Union[str, float, List[str]]]]:
-        """
-        Load all Wikidata properties with labels, descriptions, aliases, and generate embeddings for all of them.
-        
-        Returns:
-            Dictionary containing property information including aliases and their embeddings
-        """
-        sparql = SPARQLWrapper("https://query.wikidata.org/sparql")
-        sparql.addCustomHttpHeader('User-Agent', 'Bot/1.0')
-        
-        # Query to get properties with labels and aliases
-        query = """
-        SELECT ?property ?propertyLabel ?altLabel
-        WHERE {
-          ?property a wikibase:Property .
-          SERVICE wikibase:label { 
-            bd:serviceParam wikibase:language "en" .
-            ?property rdfs:label ?propertyLabel .
-          }
-          OPTIONAL { 
-            ?property skos:altLabel ?altLabel . 
-            FILTER(LANG(?altLabel) = "en")
-          }
-        }
-        """
-        
-        sparql.setQuery(query)
-        sparql.setReturnFormat(JSON)
-        sparql.setTimeout(60)
-        properties = {}
-
-        try:
-            print("🔄 Fetching Wikidata properties with aliases...")
-            results = sparql.query().convert()
-            
-            # Process results and group by property
-            property_data = {}
-            for result in results["results"]["bindings"]:
-                prop_id = result["property"]["value"].split("/")[-1]
-                label = result["propertyLabel"]["value"]
-                
-                if prop_id not in property_data:
-                    property_data[prop_id] = {
-                        "label": label,
-                        "aliases": set()
-                    }
-                
-                if "altLabel" in result:
-                    property_data[prop_id]["aliases"].add(result["altLabel"]["value"])
-            
-            print("✨ Computing embeddings for labels and aliases...")
-            # Generate embeddings for labels and aliases
-            for prop_id, data in property_data.items():
-                # Convert aliases set to list for JSON serialization
-                aliases = list(data["aliases"])
-                
-                # Generate embedding for main label
-                label_embedding = self.embedding_model.encode(data["label"]).tolist()
-                
-                # Generate embeddings for aliases
-                alias_embeddings = {}
-                for alias in aliases:
-                    alias_embedding = self.embedding_model.encode(alias).tolist()
-                    alias_embeddings[alias] = alias_embedding
-                
-                # Store all information
-                properties[prop_id] = {
-                    "label": data["label"],
-                    "label_embedding": label_embedding,
-                    "aliases": aliases,
-                    "alias_embeddings": alias_embeddings
-                }
-            
-            print(f"📝 Saving properties to {output_file}...")
-            with open(output_file, "w", encoding='utf-8') as file:
-                json.dump(properties, file, indent=4, ensure_ascii=False)
-            
-            print(f"✅ Successfully saved {len(properties)} properties with their aliases and embeddings")
-            print(f"📊 Statistics:")
-            total_aliases = sum(len(p["aliases"]) for p in properties.values())
-            print(f"   • Total properties: {len(properties)}")
-            print(f"   • Total aliases: {total_aliases}")
-            print(f"   • Average aliases per property: {total_aliases/len(properties):.2f}")
-            
-            return properties
-        
-        except Exception as e:
-            print(f"❌ Error fetching properties: {str(e)}")
-            raise
-
-    def load_embeddings(self, file_path="wikidata-properties.json"):
-        """
-        Load stored embeddings from a file and store them in a dictionary.
-        Handles both simple properties and properties with aliases.
-        """
-        try:
-            with open(file_path, "r") as file:
-                self.properties = json.load(file)
-            
-            print(f"Successfully loaded properties from {file_path}")
-            
-            # Check if we need to convert from old format to new format
-            if self.matching_config.get("use_aliases", True):
-                # Check if the first property has the new format structure
-                first_prop = next(iter(self.properties.values()))
-                if "label_embedding" not in first_prop:
-                    print("Converting properties to alias format...")
-                    converted_properties = {}
-                    for prop_id, prop_data in self.properties.items():
-                        converted_properties[prop_id] = {
-                            "label": prop_data["label"],
-                            "label_embedding": prop_data["embedding"],
-                            "aliases": [],
-                            "alias_embeddings": {}
-                        }
-                    self.properties = converted_properties
-                    print("Properties converted to alias format.")
-            else:
-                # Check if we need to convert from new format to old format
-                first_prop = next(iter(self.properties.values()))
-                if "embedding" not in first_prop:
-                    print("Converting properties to simple format...")
-                    converted_properties = {}
-                    for prop_id, prop_data in self.properties.items():
-                        converted_properties[prop_id] = {
-                            "label": prop_data["label"],
-                            "embedding": prop_data["label_embedding"]
-                        }
-                    self.properties = converted_properties
-                    print("Properties converted to simple format.")
-            
-            return self.properties
-        except Exception as e:
-            print(f"Error loading embeddings: {e}")
-            return {}
-        
-    def save_property_matches(self, predicate: str, matches: List[Dict], output_dir: str = "property_matches"):
-        """
-        Save top property matches to a file.
-        
-        Args:
-            predicate: The predicate that was searched for
-            matches: List of dictionaries containing match information
-            output_dir: Directory to save the matches file
-        """
-        try:
-            # Create output directory if it doesn't exist
-            os.makedirs(output_dir, exist_ok=True)
-            
-            # Create a filename from the predicate
-            safe_filename = quote(predicate.replace(" ", "_"))
-            output_file = os.path.join(output_dir, f"{safe_filename}_matches.txt")
-            
-            with open(output_file, 'w', encoding='utf-8') as f:
-                f.write(f"Top matches for predicate: \"{predicate}\"\n")
-                f.write("="*50 + "\n\n")
-                
-                for i, match in enumerate(matches, 1):
-                    f.write(f"{i}. Match Details:\n")
-                    f.write(f"   Property ID: {match['property_id']}\n")
-                    f.write(f"   Label: {match['label']}\n")
-                    f.write(f"   Similarity Score: {match['similarity']:.4f}\n")
-                    f.write("\n")
-                    
-            print(f"✓ Saved property matches to: {output_file}")
-            
-        except Exception as e:
-            print(f"⚠️  Error saving property matches: {str(e)}")
-
-    def find_best_match(self, predicate: str):
-        """
-        Find and save top 10 matches for a predicate based on cosine similarity.
-        Uses either simple matching or matching with aliases based on configuration.
-        """
-        if not self.properties:
-            print("No properties loaded. Please ensure properties are initialized.")
-            return None
-
-        if self.matching_config.get("use_aliases", True):
-            return self._find_best_match_with_aliases(predicate)
-        else:
-            return self._find_best_match_simple(predicate)
-
-    def _find_best_match_simple(self, predicate: str):
-        """Simple matching using only main labels."""
-        predicate_embedding = self.embedding_model.encode(predicate)
-        matches = []
-        
-        for prop_id, prop_data in self.properties.items():
-            similarity = util.cos_sim(predicate_embedding, torch.tensor(prop_data["embedding"])).item()
-            matches.append({
-                "property_id": prop_id,
-                "label": prop_data["label"],
-                "similarity": similarity
-            })
-        
-        matches.sort(key=lambda x: x['similarity'], reverse=True)
-        top_matches = matches[:10]
-        self.save_property_matches(predicate, top_matches)
-        
-        if top_matches:
-            best_match = top_matches[0]
-            print(f"\n🎯 Best Match for '{predicate}':")
-            print(f"🔑 Property ID: {best_match['property_id']}")
-            print(f"📝 Label: {best_match['label']}")
-            print(f"💯 Similarity: {best_match['similarity']:.4f}")
-            return best_match["property_id"]
-        return None
-
-    def _find_best_match_with_aliases(self, predicate: str):
-        """Advanced matching using both labels and aliases."""
-        predicate_embedding = self.embedding_model.encode(predicate)
-        matches = []
-        
-        for prop_id, prop_data in self.properties.items():
-            label_similarity = util.cos_sim(predicate_embedding, torch.tensor(prop_data["label_embedding"])).item()
-            
-            alias_similarities = []
-            for alias in prop_data["aliases"]:
-                alias_embedding = prop_data["alias_embeddings"][alias]
-                similarity = util.cos_sim(predicate_embedding, torch.tensor(alias_embedding)).item()
-                alias_similarities.append(similarity)
-            
-            best_similarity = max([label_similarity] + alias_similarities) if alias_similarities else label_similarity
-            best_match_text = prop_data["label"]
-            
-            if alias_similarities:
-                max_alias_idx = alias_similarities.index(max(alias_similarities))
-                if alias_similarities[max_alias_idx] > label_similarity:
-                    best_match_text = prop_data["aliases"][max_alias_idx]
-            
-            matches.append({
-                "property_id": prop_id,
-                "label": prop_data["label"],
-                "best_match_text": best_match_text,
-                "similarity": best_similarity
-            })
-        
-        matches.sort(key=lambda x: x['similarity'], reverse=True)
-        top_matches = matches[:10]
-        self.save_property_matches(predicate, top_matches)
-        
-        if top_matches:
-            best_match = top_matches[0]
-            print(f"\n🎯 Best Match for '{predicate}':")
-            print(f"🔑 Property ID: {best_match['property_id']}")
-            print(f"📝 Label: {best_match['label']}")
-            if best_match['best_match_text'] != best_match['label']:
-                print(f"✨ Matched via alias: {best_match['best_match_text']}")
-            print(f"💯 Similarity: {best_match['similarity']:.4f}")
-            return best_match["property_id"]
-        return None
 
     def query_wikidata_entity(self, label: str, language: str = "en") -> str:
         """
@@ -428,33 +152,28 @@ class Integrator:
             return None
             
         except Exception as e:
+
             print(f"Error searching for entity {label}: {e}")
 
     def query_wikidata_property(self, predicate: str, method: str = "sparql", file_path: str = "wikidata-properties-with-aliases.json", language: str = "en") -> str:
         """
-        Sucht nach Wikidata-Properties entweder über SPARQL oder die Wikidata-API.
+        Search for Wikidata properties using either SPARQL or the Wikidata API.
         
         Args:
-            predicate: Der zu suchende Prädikat-Text
-            method: "sparql" oder "api" für die Suchmethode
-            file_path: Pfad zur Property-Cache-Datei (nur für SPARQL)
-            language: Sprachcode (default: "en")
+            predicate: The predicate text to search for
+            method: "sparql" or "api" for the search method
+            file_path: Path to property cache file (only for SPARQL)
+            language: Language code (default: "en")
         
         Returns:
-            Property-ID oder None wenn nicht gefunden
+            Property ID or None if not found
         """
         if method.lower() == "sparql":
-            # Bestehende SPARQL-Implementierung
-            if not os.path.exists(file_path):
-                print(f"File '{file_path}' not found. Generating properties...")
-                self.load_wikidata_properties(output_file=file_path)
-
-            self.load_embeddings(file_path=file_path)
-            best_match = self.find_best_match(predicate)
+            best_match = self.matcher.find_best_match(predicate)
             return best_match
         
         elif method.lower() == "api":
-            # Neue API-Implementierung
+            # API implementation remains unchanged
             if predicate in self.property_cache:
                 return self.property_cache[predicate]
 
@@ -464,7 +183,7 @@ class Integrator:
                 "format": "json",
                 "language": language,
                 "search": predicate,
-                "type": "property"  # Suche nach Properties statt Items
+                "type": "property"
             }
             
             try:
@@ -473,16 +192,16 @@ class Integrator:
                 data = response.json()
                 
                 if data["search"]:
-                    # Überprüfe jedes Ergebnis auf exakte Übereinstimmungen
+                    # Check each result for exact matches
                     for result in data["search"]:
-                        # Prüfe Hauptlabel
+                        # Check main label
                         if result.get("label", "").lower() == predicate.lower():
                             property_id = result["id"]
                             self.property_cache[predicate] = property_id
                             print(f"Found exact match for '{predicate}': {property_id} ({result.get('label')})")
                             return property_id
                     
-                    # Wenn keine exakte Übereinstimmung, nimm das erste Ergebnis
+                    # If no exact match, take the first result
                     property_id = data["search"][0]["id"]
                     label = data["search"][0].get("label", "")
                     self.property_cache[predicate] = property_id
@@ -498,9 +217,6 @@ class Integrator:
         
         else:
             raise ValueError(f"Invalid method: {method}. Use 'sparql' or 'api'.")
-
-        print(f"No match found for predicate '{predicate}'.")
-        return None
 
     def process_triple(self, triple: Tuple[str, str, str]) -> None:
         """
@@ -539,19 +255,14 @@ class Integrator:
             self.g.add((subject_uri, property_uri, Literal(obj)))
 
     def print_statistics(self, title: str = "Current Statistics") -> None:
-        """
-        Print beautiful statistics about the graph
-        """
         stats = self.get_statistics()
         
         print("\n" + "="*50)
         print(f"📊 {title}")
         print("="*50)
-        
-        # Calculate the maximum width needed for numbers
+ 
         max_num_width = max(len(str(v)) for v in stats.values())
         
-        # Print each statistic with aligned numbers
         print(f"\n🔸 Total Triples:       {stats['total_triples']:>{max_num_width}}")
         print(f"🔸 Unique Subjects:     {stats['unique_subjects']:>{max_num_width}}")
         print(f"🔸 Unique Predicates:   {stats['unique_predicates']:>{max_num_width}}")
@@ -627,28 +338,16 @@ class Integrator:
             raise
 
     def process_triples(self, triples: List[Tuple[str, str, str]]) -> None:
-        """
-        Verarbeitet eine Liste von Tripeln
-        """
         for triple in triples:
             self.process_triple(triple)
 
     def save_graph(self, output_file: str, format: str = "turtle") -> None:
-        """
-        Speichert den RDF-Graphen in einer Datei
-        """
         self.g.serialize(destination=output_file, format=format)
 
     def load_graph(self, input_file: str, format: str = "turtle") -> None:
-        """
-        Lädt einen existierenden RDF-Graphen aus einer Datei
-        """
         self.g.parse(input_file, format=format)
 
     def get_statistics(self) -> Dict:
-        """
-        Erstellt Statistiken über den Graphen
-        """
         return {
             "total_triples": len(self.g),
             "unique_subjects": len(set(self.g.subjects())),
